0x06-YOLO v8 videos/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-cors==3.0.10
boto3==1.24.12
numpy==1.22.0
opencv-python-headless
ultralytics
torchvision
torch
Pillow==9.0.1
pandas